version 2.0.4

+++ src/com.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2014-2015 Johan Henriksson.
+ * Copyright (C) 2014-2017 Johan Henriksson.
  * All rights reserved.
  *
  * This software may be modified and distributed under the terms
@@ -15,6 +15,9 @@
 #include <assert.h>
 #include <unistd.h>
 #include "config.h"
+#include "version.h"
+
+#include <QDateTime>
 
 
 
@@ -85,10 +88,9 @@ bool Resp::isResult()
         
 Com::Com()
  : m_listener(NULL)
-#ifdef ENABLE_GDB_LOG
  ,m_logFile(GDB_LOG_FILE)
-#endif
  ,m_busy(0)
+ ,m_enableLog(false)
  {
 /*
     QByteArray array = m_process.readAllStandardOutput();
@@ -98,9 +100,6 @@ Com::Com()
 */
 
 
-#ifdef ENABLE_GDB_LOG
-     m_logFile.open(QIODevice::Truncate | QIODevice::WriteOnly | QIODevice::Text);
-#endif
 
     connect(&m_process, SIGNAL(readyReadStandardOutput ()), this, SLOT(onReadyReadStandardOutput()));
 
@@ -111,17 +110,26 @@ Com::~Com()
     // Send the command to gdb to exit cleanly
     QString text = "-gdb-exit\n";
     m_process.write((const char*)text.toLatin1());
-    
-    
-    m_process.terminate();
 
-    m_process.waitForFinished();
-    
+    if(!m_process.waitForFinished(1000))
+    {
+        m_process.terminate();
+
+        m_process.waitForFinished();
+    }
+
     // Free tokens
     while(!m_freeTokens.isEmpty())
     {
         Token *token = m_freeTokens.takeFirst();
         delete token;
+    }
+
+
+    if(m_enableLog)
+    {
+        writeLogEntry("\r\n#\r\n#\r\n");
+        m_logFile.close();
     }
 }
 
@@ -780,6 +788,13 @@ Resp *Com::parseOutput()
 }
 
 
+void Com::writeLogEntry(QString logText)
+{
+    assert(m_enableLog == true);
+    m_logFile.write(logText.toUtf8());
+    m_logFile.flush();
+}
+               
 /**
  * @brief Reads output from GDB.
  */
@@ -801,14 +816,14 @@ void Com::readTokens()
             {
                 debugMsg("row:%s", stringToCStr(row));
      
-#ifdef ENABLE_GDB_LOG
-                QString logText;
-                logText = ">> ";
-                logText += row;
-                logText += "\n";
-                m_logFile.write(stringToCStr(logText), logText.length());
-                m_logFile.flush();
-#endif
+                if(m_enableLog)
+                {
+                    QString logText;
+                    logText = ">> ";
+                    logText += row;
+                    logText += "\n";
+                    writeLogEntry(logText);
+                }
 
                 QList<Token*> list;
                 char firstChar = row[0].toLatin1();
@@ -848,10 +863,14 @@ void Com::readTokens()
     }
 }
 
-    
-void Com::readFromGdb(GdbResult *m_result, Tree *m_resultData)
+
+/**
+ * @brief Reads response from GDB
+ * @return 0 on success otherwise an errorcode.
+ */
+int Com::readFromGdb(GdbResult *m_result, Tree *m_resultData)
 {
-    
+    int rc = 0;
     //debugMsg("## '%s'",stringToCStr(row));
 
     Resp *resp = NULL;
@@ -863,8 +882,16 @@ void Com::readFromGdb(GdbResult *m_result, Tree *m_resultData)
         // Parse any data received from GDB
         resp = parseOutput();
         if(resp == NULL)
-            m_process.waitForReadyRead(100);
-       
+        {
+            if(!m_process.waitForReadyRead(100))
+            {
+                QProcess::ProcessState  state = m_process.QProcess::state();
+                if(state == QProcess::NotRunning)
+                {
+                    rc = -1;
+                }
+            }
+        }
         
         while(!m_freeTokens.isEmpty())
         {
@@ -887,6 +914,8 @@ void Com::readFromGdb(GdbResult *m_result, Tree *m_resultData)
     }
     else
     {
+        *m_result = GDB_DONE;
+    
     do
     {
         
@@ -895,8 +924,19 @@ void Com::readFromGdb(GdbResult *m_result, Tree *m_resultData)
         {
             resp = parseOutput();
             if(resp == NULL)
-                m_process.waitForReadyRead(100);
-        }while(resp == NULL);
+            {
+                if(!m_process.waitForReadyRead(100))
+                {
+                    QProcess::ProcessState  state = m_process.QProcess::state();
+                    if(state == QProcess::NotRunning)
+                    {
+                        assert(0);
+                        
+                        rc = -1;
+                    }
+                }
+            }
+        }while(resp == NULL && rc == 0);
        
         
         while(!m_freeTokens.isEmpty())
@@ -906,10 +946,11 @@ void Com::readFromGdb(GdbResult *m_result, Tree *m_resultData)
         }
 
 
-        m_respQueue.push_back(resp);
+        if(resp != NULL)
+            m_respQueue.push_back(resp);
 
         
-        if(resp->getType() == Resp::RESULT)
+        if(resp != NULL && resp->getType() == Resp::RESULT)
         {
             assert(m_resultData != NULL);
             
@@ -918,7 +959,10 @@ void Com::readFromGdb(GdbResult *m_result, Tree *m_resultData)
             m_resultData->copy(resp->tree);
         }
 
-    }while(m_result != NULL && resp->getType() != Resp::TERMINATION);
+    }while(m_result != NULL && resp != NULL && resp->getType() != Resp::TERMINATION);
+
+    if(resp == NULL && m_result != NULL)
+        *m_result = GDB_ERROR;
 }
 
     //  debugMsg("# ---<< \n");
@@ -942,13 +986,16 @@ void Com::readFromGdb(GdbResult *m_result, Tree *m_resultData)
             if(!row.isEmpty())
                 debugMsg("GDB|E>%s", stringToCStr(row));
         }
-    } 
+    }
+
+    return rc;
 }
 
 
 GdbResult Com::command(Tree *resultData, QString text)
 {
     Tree resultDataNull;
+    int rc = 0;
 
     assert(m_busy == 0);
     
@@ -975,21 +1022,24 @@ GdbResult Com::command(Tree *resultData, QString text)
     text += "\n";
     m_process.write((const char*)text.toLatin1());
 
-#ifdef ENABLE_GDB_LOG
-    //
-    QString logText;
-    logText = "\n<< ";
-    logText += text + "\n";
-    m_logFile.write(stringToCStr(logText), logText.length());
-    m_logFile.flush();
-#endif
 
+    if(m_enableLog)
+    {
+        //
+        QString logText;
+        logText = "\n<< ";
+        logText += text + "\n";
+        writeLogEntry(logText);
+    }
+    
 
     do
     {
-        readFromGdb(&result,resultData);
-
-    }while(!m_pending.isEmpty());
+        if(readFromGdb(&result,resultData))
+        {
+                rc = -1;
+        }
+    }while(!m_pending.isEmpty() && rc == 0);
 
     
     while(!m_list.isEmpty())
@@ -1004,21 +1054,39 @@ GdbResult Com::command(Tree *resultData, QString text)
     dispatchResp();
 
     onReadyReadStandardOutput();
-    
+
+    if(rc)
+        return GDB_ERROR;
     return result;
 }
 
 
-int Com::init(QString gdbPath)
+/**
+ * @brief Starts gdb
+ * @return 0 on success and gdb was started.
+ */
+int Com::init(QString gdbPath, bool enableDebugLog)
 {
     QString commandLine;
+
+    enableLog(enableDebugLog);
+
+        
     commandLine.sprintf("%s --interpreter=mi2", stringToCStr(gdbPath));
-    
-    m_process.start(commandLine);//gdb ./testapp/test");
-    m_process.waitForStarted();
+
+    if(m_enableLog)
+    {
+        QString logStr;
+        logStr = "# Gdb commandline: '" + commandLine + "'\r\n";
+        writeLogEntry(logStr);
+    }
+
+    m_process.start(commandLine);
+    m_process.waitForStarted(6000);
 
     if(m_process.state() == QProcess::NotRunning)
     {
+        assert(0);
         return 1;
     }
 
@@ -1089,4 +1157,38 @@ void Com::dispatchResp()
 
 }
 
+void Com::enableLog(bool enable)
+{
+    if(m_enableLog == enable)
+        return;
 
+    if(m_enableLog)
+    {
+        writeLogEntry("#\r\n");
+        m_logFile.close();
+    }
+    m_enableLog = false;
+
+    if(enable)
+    {
+        if(m_logFile.open(QIODevice::Truncate | QIODevice::WriteOnly | QIODevice::Text))
+        {
+            warnMsg("Created %s", (const char*)GDB_LOG_FILE);
+
+            m_enableLog = true;
+
+            QString logStr;
+            QDateTime now = QDateTime::currentDateTime();
+            logStr = "# Created: " + now.toString("yyyy-MM-dd hh:mm:ss") + "\r\n";
+            writeLogEntry(logStr);
+            logStr.sprintf("# Gede version: %d.%d.%d \r\n", GD_MAJOR,GD_MINOR,GD_PATCH);
+            writeLogEntry(logStr);
+
+        }
+        else
+            warnMsg("Failed to create log file %s", (const char*)GDB_LOG_FILE);
+        
+    }
+}
+
+
